requirement.txt
Django==4.1.4
Pillow==9.3.0
psycopg2==2.9.5
djangocms-admin-style==3.2.0
django-ckeditor==6.5.1